+++ docs/README.md
@@ -1,0 +1,2 @@
+## 리팩토링 목록
+- [x] 공통 상위 url 별로 컨트롤러 나누기
