+++ src/main/java/roomescape/controller/ReservationsController.java
@@ -1,51 +1,34 @@
-package roomescape;
+package roomescape.controller;
 
-import java.lang.reflect.Member;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 import java.util.concurrent.atomic.AtomicLong;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.domain.Reservation;
 import roomescape.dto.ReservationRequest;
 
-@Controller
-public class RoomescapeController {
+@RestController
+public class ReservationsController {
 
     private final List<Reservation> reservations = new ArrayList<>();
     private final AtomicLong index = new AtomicLong(1);
 
-    @GetMapping("/admin")
-    public String admin() {
-        return "admin/index";
-    }
-
-    @GetMapping("/admin/reservation")
-    public String reservation() {
-        return "admin/reservation-legacy";
-    }
-
     @GetMapping("/reservations")
-    @ResponseBody()
     @ResponseStatus(HttpStatus.OK)
-    public List<Reservation> reservations() {
+    public List<Reservation> read() {
         return reservations;
     }
 
     @PostMapping("/reservations")
-    @ResponseBody()
     @ResponseStatus(HttpStatus.OK)
-    public Reservation addReservation(@RequestBody ReservationRequest reservationRequest) {
+    public Reservation add(@RequestBody final ReservationRequest reservationRequest) {
         Long id = index.getAndIncrement();
         Reservation reservation = reservationRequest.toReservation(id);
         reservations.add(reservation);
@@ -53,14 +36,13 @@ public class RoomescapeController {
     }
 
     @DeleteMapping("/reservations/{id}")
-    public ResponseEntity<Void> deleteReservation(@PathVariable("id") Long id) {
+    @ResponseStatus(HttpStatus.OK)
+    public void delete(@PathVariable(name = "id") final Long id) {
         Reservation reservation = reservations.stream()
-                .filter(it -> Objects.equals(it.getId(), id))
+                .filter(it -> it.getId().equals(id))
                 .findFirst()
-                .orElseThrow(RuntimeException::new);
+                .orElseThrow(IllegalArgumentException::new);
 
         reservations.remove(reservation);
-
-        return ResponseEntity.ok().build();
     }
 }
