+++ src/main/java/roomescape/dto/ReservationRequest.java
@@ -2,7 +2,7 @@ package roomescape.dto;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
-import roomescape.Reservation;
+import roomescape.domain.Reservation;
 
 public record ReservationRequest(String name, LocalDate date, LocalTime time) {
